grocery-bill-ai/requirements.txt
Flask
flask-cors
pytesseract
Pillow
pdf2image
pandas
easyocr
numpy